requirements.txt
apiview==1.3.25
asgiref==3.8.1
diff-match-patch==20230430
Django==5.1.1
django-environ==0.11.2
django-import-export==4.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1
tablib==3.5.0
